Tests: Move schemas unit tests into unit test class
Running the `SQLExecutor` based unit tests against a `clusterService`
instance setup via the integration test setup can cause issues. It is
messing with the cluster state.
